+++ projects/minecraft/v1.19.2/changes.md
@@ -3,3 +3,7 @@
 ## Pets / Tameable (e.g. Dogs, Cats, Horses)
 
 - Death is prevented when tamed.
+
+## Latency
+
+- Allows Elytra flight to start and stop client-side.